clean up textgen script
